+++ docs/make.jl
@@ -24,9 +24,5 @@ makedocs(
 )
 
 deploydocs(
-    deps = nothing,
-    make = nothing,
-    target = "build",
     repo = "github.com/lanl-ansi/ThreePhasePowerModels.jl.git",
-    julia = "0.6"
 )