+++ src/io/dss_parse.jl
@@ -5,7 +5,7 @@ end
 
 
 double_operators = Dict("+" => +, "-" => -, "*" => *, "/" => /, "^" => ^,
-                        "atan2" => f(x, y) = rad2deg(atan2(y, x)))
+                        "atan2" => (x, y) -> rad2deg(atan2(y, x)))
 
 single_operators = Dict("sqr" => sqr, "sqrt" => sqrt, "inv" => inv, "ln" => log,
                         "exp" => exp, "log10" => log10, "sin" => sind, "cos" => cosd,
@@ -17,13 +17,13 @@ array_delimiters = ['\"', '\'', '[', '{', '(', ']', '}', ')']
 function parse_rpn(expr::AbstractString, dtype::Type=Float64)
     clean_expr = strip(expr, array_delimiters)
 
-    if contains(clean_expr, "rollup") || contains(clean_expr, "rolldn") || contains(clean_expr, "swap")
+    if occursin("rollup", clean_expr) || occursin("rolldn", clean_expr) || occursin("swap", clean_expr)
         warn(LOGGER, "parse_rpn does not support \"rollup\", \"rolldn\", or \"swap\", leaving as String")
         return expr
     end
 
     stack = []
-    split_expr = contains(clean_expr, ",") ? split(clean_expr, ',') : split(clean_expr)
+    split_expr = occursin(",", clean_expr) ? split(clean_expr, ',') : split(clean_expr)
 
     for item in split_expr
         try
@@ -329,7 +329,7 @@ end
 "parse matrices according to active nodes"
 function parse_matrix(data::Array{T}, nodes::Array{Bool}, nph::Int=3, fill_val=0.0)::Array where T
     mat = fill(fill_val, (nph, nph))
-    idxs = find(nodes[1:nph])
+    idxs = findall(nodes[1:nph])
 
     for i in 1:size(idxs)[1]
         mat[idxs[i], idxs[i]] = data[i, i]
@@ -344,7 +344,7 @@ end
 
 "checks if `data` is an opendss-style matrix string"
 function isa_matrix(data::AbstractString)::Bool
-    if contains(data, "|")
+    if occursin("|", data)
         return true
     else
         return false
@@ -360,14 +360,15 @@ Parses a OpenDSS style array string `data` into a one dimensional array of type
 quotes, and elements are separated by spaces.
 """
 function parse_array(dtype::Type, data::AbstractString)
-    if contains(data, ",")
+    if occursin(",", data)
         split_char = ','
     else
         split_char = ' '
     end
 
     if isa_rpn(data)
-        if length(matchall(Regex(string("[",join(array_delimiters, '\\'),"]")),data)) == 2
+        matches = collect((m.match for m = eachmatch(Regex(string("[",join(array_delimiters, '\\'),"]")), data, overlap=false)))
+        if length(matches) == 2
             if dtype == String
                 return data
             else
@@ -405,7 +406,7 @@ end
 "parse matrices according to active nodes"
 function parse_array(data, nodes::Array{Bool}, nph::Int=3, fill_val=0.0)::Array
     mat = fill(fill_val, nph)
-    idxs = find(nodes[1:nph])
+    idxs = findall(nodes[1:nph])
 
     if length(data) == 1 && nph > 1
         for i in idxs
@@ -424,8 +425,8 @@ end
 "checks if `data` is an opendss-style array string"
 function isa_array(data::AbstractString)::Bool
     clean_data = strip(data)
-    if !contains(clean_data, "|")
-        if contains(clean_data, ",")
+    if !occursin("|", clean_data)
+        if occursin(",", clean_data)
             return true
         elseif startswith(clean_data, "[") && endswith(clean_data, "]")
             return true
@@ -469,7 +470,7 @@ Parses a Bus Coordinate `file`, in either "dat" or "csv" formats, where in
 to contain "bus,x,y" on each line.
 """
 function parse_buscoords(file::AbstractString)::Array
-    file_str = readstring(open(file))
+    file_str = read(open(file), String)
     regex = r",\s*"
     if endswith(lowercase(file), "csv") || endswith(lowercase(file), "dss")
         regex = r","
@@ -505,14 +506,14 @@ function parse_properties(properties::AbstractString)::Array
     endDQuot = true
     str_out = ""
 
-    properties = replace(properties, r"\s*=\s*", "=")
+    properties = replace(properties, r"\s*=\s*" => "=")
     nchars = length(properties)
 
     for (n, char) in enumerate(properties)
         sstr_out = split(str_out, "=")
         if length(sstr_out) == 2 && sstr_out[2] != ""
             endEquality = true
-        elseif !contains(str_out, "=") && (char == ' ' || n == nchars)
+        elseif !occursin("=", str_out) && (char == ' ' || n == nchars)
             endEquality = true
         else
             endEquality = false
@@ -579,9 +580,9 @@ new value is appended to the end.
 function add_property(compDict::Dict, key::AbstractString, value::Any)::Dict
     if haskey(compDict, lowercase(key))
         rmatch = match(r"_(\d+)$", key)
-        if typeof(rmatch) != Void
+        if typeof(rmatch) != Nothing
             endNum = parse(Int, rmatch.captures[1]) + 1
-            key = replace(key, r"_(\d+)$", "_$endNum")
+            key = replace(key, r"_(\d+)$" => "_$endNum")
         else
             key = string(key, "_2")
         end
@@ -617,34 +618,34 @@ function parse_component(component::AbstractString, properties::AbstractString, 
     for (n, property) in enumerate(propArray)
         if property == ""
             continue
-        elseif !contains(property, "=")
+        elseif !occursin("=", property)
             property = join([propNames[propIdx], property], '=')
             propIdx += 1
         else
             if split(component,'.')[1] == "loadshape" && startswith(property, "mult")
-                property = replace(property, "mult", "pmult")
+                property = replace(property, "mult" => "pmult")
             end
 
             try
-                propIdxs = find(e->e==split(property,'=')[1], propNames)
+                propIdxs = findall(e->e==split(property,'=')[1], propNames)
                 if length(propIdxs) > 0
-                    propIdx = find(e->e==split(property,'=')[1], propNames)[1] + 1
+                    propIdx = findall(e->e==split(property,'=')[1], propNames)[1] + 1
                 end
             catch e
                 if split(component,'.')[1] == "transformer"
                     if split(property,'=')[1] == "ppm"
-                        property = replace(property, "ppm", "ppm_antifloat")
+                        property = replace(property, "ppm" => "ppm_antifloat")
                     elseif split(property,'=')[1] == "x12"
-                        property = replace(property, "x12", "xhl")
+                        property = replace(property, "x12" => "xhl")
                     elseif split(property,'=')[1] == "x23"
-                        property = replace(property, "x23", "xlt")
+                        property = replace(property, "x23" => "xlt")
                     elseif split(property,'=')[1] == "x13"
-                        property = replace(property, "x13", "xht")
+                        property = replace(property, "x13" => "xht")
                     end
                 else
                     throw(e)
                 end
-                propIdx = find(e->e==split(property,'=')[1], propNames)[1] + 1
+                propIdx = findall(e->e==split(property,'=')[1], propNames)[1] + 1
             end
         end
 
@@ -739,7 +740,7 @@ function parse_dss(io::IOStream)::Dict
     info(LOGGER, "Calling parse_dss on $filename")
     currentFile = split(filename, "/")[end]
     path = join(split(filename, '/')[1:end-1], '/')
-    dss_str = readstring(open(filename))
+    dss_str = read(open(filename), String)
     dss_data = Dict{String,Array}()
 
     dss_data["filename"] = [currentFile]
@@ -753,7 +754,7 @@ function parse_dss(io::IOStream)::Dict
     nlines = length(stripped_lines)
 
     for (n, line) in enumerate(stripped_lines)
-        real_line_num = find(lines .== line)[1]
+        real_line_num = findall(lines .== line)[1]
         debug(LOGGER, "LINE $real_line_num: $line")
         line = strip_comments(line)
 
@@ -938,13 +939,13 @@ function parse_busname(busname::AbstractString)
 
     nodes = Array{Bool}([0 0 0 0])
 
-    for num in range(1,3)
-        if contains(elements, "$num")
+    for num in 1:3
+        if occursin("$num", elements)
             nodes[num] = true
         end
     end
 
-    if contains(elements, "0") || sum(nodes[1:3]) == 1
+    if occursin("0", elements) || sum(nodes[1:3]) == 1
         nodes[4] = true
     end