+++ test/runtests.jl
@@ -3,6 +3,8 @@ const TPPMs = ThreePhasePowerModels
 
 using Memento
 
+using InfrastructureModels
+
 using PowerModels
 const PMs = PowerModels
 
@@ -15,8 +17,16 @@ using Pavito
 using Juniper
 using SCS
 
-using Base.Test
+using Compat.Test
+using Compat.LinearAlgebra
 
+if VERSION < v"0.7.0-"
+    pms_path = Pkg.dir("PowerModels")
+end
+
+if VERSION > v"0.7.0-"
+    pms_path = joinpath(dirname(pathof(PowerModels)), "..")
+end
 
 ipopt_solver = IpoptSolver(tol=1e-6, print_level=0)
 cbc_solver = CbcSolver()