+++ test/tp_opf.jl
@@ -1,7 +1,7 @@
 
 @testset "test make multi-phase" begin
     @testset "3-bus 3-phase case" begin
-        mp_data = PMs.parse_file("../../PowerModels/test/data/matpower/case3.m")
+        mp_data = PMs.parse_file("$(pms_path)/test/data/matpower/case3.m")
         PMs.make_multiconductor(mp_data, 3)
         result = run_tp_opf(mp_data, PMs.ACPPowerModel, ipopt_solver)