+++ test/tp_opf_bf.jl
@@ -1,6 +1,6 @@
 @testset "test linearised distflow opf_bf" begin
     @testset "3-bus case" begin
-        mp_data = PowerModels.parse_file("../../PowerModels/test/data/matpower/case3.m")
+        mp_data = PowerModels.parse_file("$(pms_path)/test/data/matpower/case3.m")
         PowerModels.make_multiconductor(mp_data, 3)
         result = run_tp_opf_bf(mp_data, LPLinUBFPowerModel, ipopt_solver)
 
@@ -39,7 +39,7 @@ end
 
 @testset "test linearised distflow opf_bf in diagonal matrix form" begin
     @testset "3-bus case" begin
-        mp_data = PowerModels.parse_file("../../PowerModels/test/data/matpower/case3.m")
+        mp_data = PowerModels.parse_file("$(pms_path)/test/data/matpower/case3.m")
         PowerModels.make_multiconductor(mp_data, 3)
         result = run_tp_opf_bf(mp_data, LPdiagUBFPowerModel, ipopt_solver)
 
@@ -79,7 +79,7 @@ end
 
 @testset "test linearised distflow opf_bf in full matrix form" begin
     @testset "3-bus case" begin
-        mp_data = PowerModels.parse_file("../../PowerModels/test/data/matpower/case3.m")
+        mp_data = PowerModels.parse_file("$(pms_path)/test/data/matpower/case3.m")
         PowerModels.make_multiconductor(mp_data, 3)
         result = run_tp_opf_bf(mp_data, LPfullUBFPowerModel, ipopt_solver)