+++ CHANGELOG.md
@@ -12,6 +12,7 @@ ThreePhasePowerModels.jl Change Log
 - Fixed parsing OpenDSS files containing redirect/compile/buscoords on case-sensitive filesystems
 - Add 'source_id' field to components parsed from OpenDSS, to help determine origin and active phases
 - Add conversion of OpenDSS PVSystem objects into generators, using KVA for generator limits
+- Add compatibility for Julia v0.7/v1.0
 
 ### v0.1.2
 - Add support for network flow approximation formulation, NFAPowerModel